# docs/swagger.yaml
basePath: /api
definitions:
...
  models.InvoiceItemInput:
    properties:
      description:
        type: string
      quantity:
        minimum: 0
        type: number
      rate:
        type: number
    required:
    - description
    type: object
...
  models.SubscriptionInput:
    properties:
      cadence:
        enum:
        - weekly
        - monthly
        type: string
      client_id:
        type: string
      currency:
        maxLength: 3
        type: string
      discount:
        type: number
      due_days:
        maximum: 365
        minimum: 0
        type: integer
      invoice_status:
        enum:
        - draft
        - sent
        type: string
      items:
        items:
          $ref: '#/definitions/models.InvoiceItemInput'
        minItems: 1
        type: array
      lead_days:
        maximum: 90
        minimum: 0
        type: integer
      name:
        maxLength: 120
        type: string
      notes:
        type: string
      payment_method:
        description: PaymentMethod mirrors the Invoice column; see the comment there.
        enum:
        - Cash
        - Bank transfer
        - Online
        maxLength: 32
        type: string
      start_date:
        type: string
      status:
        enum:
        - active
        - paused
        type: string
      tax_rate:
        minimum: 0
        type: number
      terms:
        type: string
    required:
    - cadence
    - currency
    - invoice_status
    - items
    - name
    - status
    type: object
  models.SubscriptionStatusInput:
    properties:
      status:
        enum:
        - active
        - paused
        type: string
    required:
    - status
    type: object
...
info:
  contact:
    name: API Support
  description: |-
    TuPay API. Two prefixes serve the same routes: /api/v1
    (current, used by the MPA) and /api (legacy, deprecated —
    responses carry a Sunset header). Unless tagged otherwise,
    endpoints speak JSON with this envelope. Success: the data
    keys directly, e.g. {"expense": {...}} or
    {"invoices": [...], "meta": {"page": 1, "per_page": 20, "total": 42}}.
    List endpoints accept ?page (default 1) and ?per_page
    (default 20, max 100) and always return a meta object.
    Errors: {"error": {"message": "...", "details": {...}}} with a
    stable English message (translated client-side via i18n api.*
    keys). Logout answers 204 with no body; most DELETEs do too,
    except DELETE /payments/:id which voids and answers 200.

    Auth — three independent schemes, never mixed:
    1. SessionCookie (most endpoints): login/register set
    HttpOnly access_token + refresh_token cookies. Send cookies
    with every request (fetch: credentials:include). An expired
    access token is refreshed transparently from the refresh
    cookie. A Bearer access token in the Authorization header
    works as a fallback where cookies are unavailable.
    2. CSRF double-submit (session-cookie POST/PATCH/DELETE only):
    login/register also set a readable csrf_token cookie — echo
    it back as the X-CSRF-Token header or the mutation is
    rejected with 403. Public GETs and API-key calls are exempt.
...
  license:
...
paths:
...
  /subscriptions:
    get:
      description: Get subscriptions of current org.
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - SessionCookie: []
      summary: list subscriptions
      tags:
      - Subscriptions
    post:
      consumes:
      - application/json
      description: Create a subscription.
      parameters:
      - description: Subscription payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SubscriptionInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - SessionCookie: []
      summary: create subscription
      tags:
      - Subscriptions
  /subscriptions/{id}:
    delete:
      description: Delete a subscription.
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ok
          schema:
            type: string
      security:
      - SessionCookie: []
      summary: delete subscription
      tags:
      - Subscriptions
    get:
      description: Get a subscription by ID.
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - SessionCookie: []
      summary: get subscription
      tags:
      - Subscriptions
    patch:
      consumes:
      - application/json
      description: Update a subscription.
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Subscription payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SubscriptionInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - SessionCookie: []
      summary: update subscription
      tags:
      - Subscriptions
  /subscriptions/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update a subscription's status.
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Status payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SubscriptionStatusInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - SessionCookie: []
      summary: update subscription status
      tags:
      - Subscriptions
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  SessionCookie:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"